Fix wrong function name in model interpretation
I was getting an error about the `inpackage` function not existing, it turns out there was a typo, as the function is called `_inpackage`. 

I found this by trying out the model interpretation functionality by passing a CatBoost model, without supplying my own `model_predict_function`.  So the error popped up in the 3rd if statement.